refactor(examples): make example apps framework-oriented
The initial example apps were named and organized around Stripe, which made the app boundary look provider-specific even though future providers should fit into the same framework examples. This restructures the examples around Express and Next.js while keeping each Webhook Endpoint provider-specific.

Stripe sender tooling now lives under provider-namespaced scripts, endpoint creation is kept at the route boundary, and provider-specific configuration sits beside the Stripe route so future provider examples follow the same shape without sharing hidden helper modules.
