+++ apps/examples/nextjs/eslint.config.js
