+++ apps/examples/nextjs/next.config.mjs
