+++ apps/examples/nextjs/src/telemetry.ts
