+++ turbo.json
@@ -11,12 +11,12 @@
     "DATABASE_URL",
     "NODE_ENV",
     "PORT",
-    "BETTER_WEBHOOK_ENDPOINT_IDENTITY",
-    "BETTER_WEBHOOK_IDEMPOTENCY_TTL_MS",
     "BETTER_WEBHOOK_OTEL_ENDPOINT",
-    "BETTER_WEBHOOK_REPLAY_WINDOW_MS",
-    "BETTER_WEBHOOK_URL",
-    "OTEL_SERVICE_NAME"
+    "OTEL_SERVICE_NAME",
+    "STRIPE_WEBHOOK_ENDPOINT_IDENTITY",
+    "STRIPE_WEBHOOK_IDEMPOTENCY_TTL_MS",
+    "STRIPE_WEBHOOK_REPLAY_WINDOW_MS",
+    "STRIPE_WEBHOOK_URL"
   ],
   "tasks": {
     "build": {
